% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnrhash.R
\name{FL_autodecoder_logic}
\alias{FL_autodecoder_logic}
\alias{FL_autodecoder_ui}
\alias{FL_ui}
\alias{FL_encoder_ui}
\title{Custom learnr methods}
\usage{
FL_autodecoder_logic(question_ids = NULL)

FL_autodecoder_ui()

FL_ui(url, add_summary = TRUE, before = TRUE)

FL_encoder_ui(url, add_summary = TRUE)
}
\arguments{
\item{question_ids}{character vector}

\item{url}{character}

\item{add_summary}{logical}

\item{before}{logical; If \code{TRUE}, generates the ui to place before the hash
generator. Otherwise generates the submission form ui, placed after the
generator.}
}
\description{
Custom learnr methods
}
