% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnrhash.R
\name{grade_alternatives}
\alias{grade_alternatives}
\title{Grade alternatives}
\usage{
grade_alternatives(valid_answers, envir = parent.frame())
}
\arguments{
\item{valid_answers}{A vector of valid answers}

\item{envir}{grading environment}
}
\description{
Grade alternatives
}
