% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnrhash.R
\name{grade_alternatives}
\alias{grade_alternatives}
\title{Grade alternatives}
\usage{
grade_alternatives(valid_answers, .result, multi = FALSE)
}
\arguments{
\item{valid_answers}{A vector of valid answers}

\item{.result}{The result to be checked for valid alternative}

\item{multi}{logical; If TRUE, allow vector results. Default: FALSE}
}
\description{
Grade alternatives
}
